Contract Audit for OriginCash
We have audited the contract on https://solidityscan.com and achieved a fairly high score. There are no high-risk vulnerabilities that could lead to any losses for you. On the other hand, it couldn't achieve a perfect score because we have designed a unique system that breaks the link between asset inputs and outputs.